+++ sw/logger/txtwrite.cpp
@@ -17,7 +17,7 @@ void txtwrite::txtwrite_state(ofstream &logfile)
   t << simtime_seconds;
   vector<Agent*> state_buff = s;
 
-  for (int i = 0; i < nagents; i++)
+  for (uint8_t i = 0; i < nagents; i++)
   {
     logfile << i+1 << " ";
     logfile << t.str() << " "; // divide by 1000000.0
